+++ src/Airbnb.AppService/Responses/Listings/UpdateListingPriceResponse.cs
@@ -1,0 +1,8 @@
+﻿using Airbnb.SharedKernel.Responses;
+
+namespace Airbnb.AppService.Responses.Listings;
+
+public class UpdateListingPriceResponse : BaseResponse
+{
+    
+}