+++ src/Airbnb.AppService/Validations/Listings/UpdateListingPriceValidator.cs
@@ -1,0 +1,11 @@
+﻿using Airbnb.AppService.Commands.Listings.UpdateListingPrice;
+using FluentValidation;
+
+namespace Airbnb.AppService.Validations.Listings;
+
+public class UpdateListingPriceValidator : AbstractValidator<UpdateListingPriceCommand>
+{
+    public UpdateListingPriceValidator()
+    {
+    }
+}