+++ src/Airbnb.Core/Events/Events/Listings/ListingPriceRollback/ListingPriceRollbackEvent.cs
@@ -1,0 +1,9 @@
+﻿using Airbnb.SharedKernel.Events;
+
+namespace Airbnb.Core.Events.Events.Listings.ListingPriceRollback;
+
+public class ListingPriceRollbackEvent : IDomainEvent
+{
+    public Guid ListingId { get; set; }
+    public decimal CurrentPrice { get; set; }
+}