+++ src/Airbnb.Core/Events/Events/Listings/ListingPriceUpdated/ListingPriceUpdatedEvent.cs
@@ -1,6 +1,9 @@
-﻿namespace Airbnb.Core.Events.Events.Listings.ListingPriceUpdated;
+﻿using Airbnb.SharedKernel.Events;
 
-public class ListingPriceUpdatedEvent
+namespace Airbnb.Core.Events.Events.Listings.ListingPriceUpdated;
+
+public class ListingPriceUpdatedEvent : IDomainEvent
 {
-
+    public Guid ListingId { get; set; }
+    public decimal NewPrice { get; set; }
 }