+++ src/Airbnb.Core/Repositories/IListingRepository.cs
@@ -5,4 +5,6 @@ namespace Airbnb.Core.Repositories;
 public interface IListingRepository
 {
     Task AddAsync(Listing listing);
+    Task<Listing?> GetByIdAsync(Guid listingId);
+    Task UpdateAsync(Listing listing);
 }